// frontend/src/app/api/model/medicine-response-dto.ts
/**
 * Pharmacy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MedicineResponseDto { 
    id?: number;
    name?: string;
    price?: number;
    manufacturer?: string;
    description?: string;
    prescriptionRequired?: boolean;
    quantity?: number;
}