// frontend/src/app/api/model/medicine-update-dto.ts
/**
 * Pharmacy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface MedicineUpdateDto { 
    name?: string;
    price?: number;
    manufacturer?: string;
    description?: string;
    prescriptionRequired?: boolean;
}