// frontend/src/app/api/model/inventory-add-dto.ts
/**
 * Pharmacy API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InventoryAddDto { 
    medicineId: number;
    supplierId?: number;
    quantity: number;
    batchNumber: string;
    expirationDate: string;
}